Now ACTUALLY handles blank lines and comments.  Implemented handling of 'cd' command for home (via environment) and for changing the directory (absolute or relative path) via chdir().  Implemented handling of other command via fork(), waitpid(), and execvp().  Need to implement 'status' command next.

+++ davidss2.c
@@ -2,8 +2,10 @@
 Citations:
 
 1. Command Line parser & struct taken from sample_parser.c
-    (lines 26-32, 53-77)
-2. 
+    (Lines 30-36, 91-120)
+2. Forking and waiting for child process adapted from Module 6,
+	Exploration 3 (Process API - Monitoring Child Processes).
+	(Lines 68-86)
 3.
 4.
 5.
@@ -16,6 +18,8 @@ Citations:
 #include <string.h>
 #include <signal.h>
 #include <unistd.h>
+#include <sys/types.h>
+#include <sys/wait.h>
 
 
 #define INPUT_LENGTH     2048
@@ -34,17 +38,51 @@ struct command_line {
 int main() {
 	struct command_line *curr_command;
 
-	while(true)
-	{
+	while(true) {
 		curr_command = parse_input();
 
 		// HANDLE BUILT IN COMMANDS
-		if (strcmp(curr_command->argv[0], "exit") == 0){
+		if (curr_command->argv[0] == NULL) { continue; }
+			// handle empty lines 
+		else if ((strcmp(curr_command->argv[0], "#") == 0)) { continue; }
+			// handle comments
+		else if (strcmp(curr_command->argv[0], "exit") == 0) {
 			// exit command
-			// kills any other process
+			// kills any other process via parent ID
 			pid_t groupID = getpgrp();
 			killpg(groupID, SIGKILL);
 			return EXIT_SUCCESS;
+		} else if (strcmp(curr_command->argv[0], "cd") == 0) {
+			// cd command
+			// check if cd is alone in input
+			if (curr_command->argc == 1) {
+				// go to home env
+				char * home = getenv("HOME");
+				chdir(home);
+			} else {
+				// change the directory
+				if (chdir(curr_command->argv[1]) == -1) {
+					printf("%s: no such file or directory\n", curr_command->argv[1]);
+				} else { continue; }
+			}
+		} else {
+			int childStatus;
+			pid_t pid = fork();
+			if (pid == -1) {
+				perror("fork() failed!"); 
+				exit(EXIT_FAILURE); 
+				break;
+			} else if (pid == 0) {
+			// child process executes this
+				execvp(curr_command->argv[0], curr_command->argv);
+				printf("%s: no such command\n", curr_command->argv[0]);
+				fflush(stdout);
+				exit(1);
+				break;
+			} else if (pid > 0) {
+			// parent process executes this
+				waitpid(pid, &childStatus, 0);
+			}
 		}
 	}
 	return EXIT_SUCCESS;
@@ -58,6 +96,9 @@ struct command_line *parse_input() {
 	printf(": ");
 	fflush(stdout);
 	fgets(input, INPUT_LENGTH, stdin);
+
+	// debugging print statement
+	// printf("INPUT = %s", input);
 
 	// Tokenize the input
 	char *token = strtok(input, " \n");
@@ -73,5 +114,7 @@ struct command_line *parse_input() {
 		}
 		token=strtok(NULL," \n");
 	}
+	// debugging print statement
+	// printf("CURRENT COMMAND = %s %s\n", curr_command->argv[0], curr_command->argv[1]);
 	return curr_command;
 }
